+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/detail/CardReaderDetailViewModel.kt
@@ -8,14 +8,14 @@ import com.woocommerce.android.AppPrefs
 import com.woocommerce.android.R
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
-import com.woocommerce.android.cardreader.CardReader
+import com.woocommerce.android.cardreader.connection.CardReader
 import com.woocommerce.android.cardreader.CardReaderManager
-import com.woocommerce.android.cardreader.CardReaderStatus.Connected
-import com.woocommerce.android.cardreader.SoftwareUpdateAvailability
-import com.woocommerce.android.cardreader.SoftwareUpdateAvailability.CheckForUpdatesFailed
-import com.woocommerce.android.cardreader.SoftwareUpdateAvailability.Initializing
-import com.woocommerce.android.cardreader.SoftwareUpdateAvailability.UpToDate
-import com.woocommerce.android.cardreader.SoftwareUpdateAvailability.UpdateAvailable
+import com.woocommerce.android.cardreader.connection.CardReaderStatus.Connected
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateAvailability
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateAvailability.CheckForUpdatesFailed
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateAvailability.Initializing
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateAvailability.UpToDate
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateAvailability.UpdateAvailable
 import com.woocommerce.android.extensions.exhaustive
 import com.woocommerce.android.model.UiString
 import com.woocommerce.android.model.UiString.UiStringRes
