+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/update/CardReaderUpdateViewModel.kt
@@ -10,11 +10,11 @@ import com.woocommerce.android.analytics.AnalyticsTracker.Stat.CARD_READER_SOFTW
 import com.woocommerce.android.analytics.AnalyticsTracker.Stat.CARD_READER_SOFTWARE_UPDATE_TAPPED
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
 import com.woocommerce.android.cardreader.CardReaderManager
-import com.woocommerce.android.cardreader.SoftwareUpdateStatus.Failed
-import com.woocommerce.android.cardreader.SoftwareUpdateStatus.Initializing
-import com.woocommerce.android.cardreader.SoftwareUpdateStatus.Installing
-import com.woocommerce.android.cardreader.SoftwareUpdateStatus.Success
-import com.woocommerce.android.cardreader.SoftwareUpdateStatus.UpToDate
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateStatus.Failed
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateStatus.Initializing
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateStatus.Installing
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateStatus.Success
+import com.woocommerce.android.cardreader.firmware.SoftwareUpdateStatus.UpToDate
 import com.woocommerce.android.extensions.exhaustive
 import com.woocommerce.android.model.UiString
 import com.woocommerce.android.model.UiString.UiStringRes
