+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/connection/CardReader.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.cardreader
+package com.woocommerce.android.cardreader.connection
 
 interface CardReader {
     val id: String?
