+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/connection/CardReaderDiscoveryEvents.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.cardreader
+package com.woocommerce.android.cardreader.connection
 
 sealed class CardReaderDiscoveryEvents {
     object Started : CardReaderDiscoveryEvents()
