+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/firmware/SoftwareUpdateAvailability.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.cardreader
+package com.woocommerce.android.cardreader.firmware
 
 sealed class SoftwareUpdateAvailability {
     object Initializing : SoftwareUpdateAvailability()
