+++ libs/cardreader/src/main/java/com/woocommerce/android/cardreader/payments/PaymentInfo.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.cardreader
+package com.woocommerce.android.cardreader.payments
 
 import java.math.BigDecimal
